backend/worker/requirements.txt
# Celery und Redis
celery==5.3.6
redis==4.4.4
gevent==23.9.1 # Behalten für Celery Pools

# Datenbank-Anbindung
SQLAlchemy==2.0.19
SQLAlchemy-Utils==0.41.1 # Behalten (Vorsicht)
psycopg2-binary==2.9.9
shortuuid==1.0.11 # Behalten (Vorsicht)

# OpenAI und KI-Bibliotheken
openai
backoff==2.2.1
tiktoken==0.5.2
aiohttp==3.9.1 # Behalten (Vorsicht)

# Dateiverarbeitung und Analyse
PyMuPDF==1.24.10 # Für PDF
python-docx==1.1.0 # Für DOCX
langdetect==1.0.9 # Behalten (Vorsicht)

# Allgemeine Hilfsbibliotheken
python-dotenv==1.0.1
# requests==2.32.3 # Entfernt (httpx vorhanden)
pytz==2023.3
# cryptography==41.0.3 # Entfernt (wahrscheinlich nicht benötigt)
ujson==5.8.0
colorama==0.4.6

# Protokollierung und Überwachung
psutil==5.9.5 # Für Maintenance Task
rich==13.7.0 # Für Logging
pydantic==2.5.2 # Behalten (Vorsicht)

# Dienstprogramme
# marshmallow==3.21.1 # Entfernt (wahrscheinlich nicht benötigt)
httpx==0.27.0 # Für OpenAI